REF: Restructure communication resources
