+++ octue/cli.py
@@ -6,7 +6,7 @@ import pkg_resources
 
 from octue.definitions import CHILDREN_FILENAME, FOLDER_DEFAULTS, MANIFEST_FILENAME, VALUES_FILENAME
 from octue.logging_handlers import get_remote_handler
-from octue.resources import Service, service_backends
+from octue.resources.communication import Service, service_backends
 from octue.runner import Runner
 from twined import Twine
