+++ octue/resources/__init__.py
@@ -3,7 +3,6 @@ from .child import Child
 from .datafile import Datafile
 from .dataset import Dataset
 from .manifest import Manifest
-from .service import Service
 
 
-__all__ = "Analysis", "Child", "Datafile", "Dataset", "Manifest", "Service"
+__all__ = "Analysis", "Child", "Datafile", "Dataset", "Manifest"
