+++ octue/resources/communication/__init__.py
@@ -1,0 +1,4 @@
+from .google_pub_sub.service import Service
+
+
+__all__ = ("Service",)
