+++ octue/resources/communication/google_pub_sub/service.py
@@ -7,8 +7,8 @@ from google.cloud import pubsub_v1
 
 from octue import exceptions
 from octue.mixins import CoolNameable
+from octue.resources.communication.google_pub_sub import Subscription, Topic
 from octue.resources.manifest import Manifest
-from octue.resources.pubsub import Subscription, Topic
 
 
 logger = logging.getLogger(__name__)
